+++ src/bigdata_research_tools/search/screener_search.py
@@ -1,18 +1,11 @@
-from itertools import chain
-from json import JSONDecodeError
 from logging import Logger, getLogger
-from re import findall
-from time import sleep
-from typing import List, Optional, Tuple
+from typing import List, Optional, Dict
 
-from bigdata_client.connection import RequestMaxLimitExceeds
 from bigdata_client.document import Document
 from bigdata_client.models.advanced_search_query import ListQueryComponent
 from bigdata_client.models.entities import Company
 from bigdata_client.models.search import DocumentType, SortBy
-from bigdata_client.query_type import QueryType
 from pandas import DataFrame
-from pydantic import ValidationError
 from tqdm import tqdm
 
 from bigdata_research_tools.client import bigdata_connection
@@ -26,6 +19,9 @@ from bigdata_research_tools.search.query_builder import (
     create_date_ranges,
 )
 from bigdata_research_tools.search.search import run_search
+from bigdata_research_tools.tracing import Trace, TraceEventNames, send_trace
+from bigdata_research_tools.search.search_utils import filter_search_results
+
 
 logger: Logger = getLogger(__name__)
 
@@ -39,7 +35,7 @@ def search_by_companies(
     fiscal_year: Optional[int] = None,
     sources: Optional[List[str]] = None,
     keywords: Optional[List[str]] = None,
-    control_entities: Optional[List[str]] = None,
+    control_entities: Optional[Dict] = None,
     freq: str = "M",
     sort_by: SortBy = SortBy.RELEVANCE,
     rerank_threshold: Optional[float] = None,
@@ -62,8 +58,7 @@ def search_by_companies(
         sources (Optional[List[str]]): List of sources to filter on. If none, we search across all sources.
         keywords (List[str]): A list of keywords for constructing keyword queries.
             If None, no keyword queries are created.
-        control_entities (List[str]): A list of control entity IDs for creating co-mentions queries.
-            If None, no control queries are created.
+        control_entities (Dict): A dictionary of control entities of different types for creating co-mentions queries.
         freq (str): The frequency of the date ranges. Defaults to '3M'.
         sort_by (SortBy): The sorting criterion for the search results.
             Defaults to SortBy.RELEVANCE.
@@ -99,90 +94,99 @@ def search_by_companies(
             - masked_text: str
             - other_entities_map: List[Tuple[int, str]]
     """
-    # Extract entities for search querying
-    entity_keys = [entity.id for entity in companies]
 
-   # Create entity configs
-    entities_config = EntitiesToSearch(companies=entity_keys)
-    
-    # If control_entities are provided, create a control EntityConfig
-    # For this example, assuming control_entities are all company entities
-    control_entities_config = None
-    if control_entities:
-        control_entities_config = EntitiesToSearch(companies=control_entities)
+    if not kwargs.get("current_trace"):
+        current_trace = Trace(
+            event_name=TraceEventNames.COMPANY_SEARCH,
+            document_type=scope,
+            start_date=start_date,
+            end_date=end_date,
+            rerank_threshold=rerank_threshold,
+            llm_model=None,
+            frequency=freq,
+            workflow_start_date=Trace.get_time_now(),
+        )
 
-    # Build batched queries
-    batched_query = build_batched_query(
-        sentences=sentences,
-        keywords=keywords,
-        entities=entities_config,
-        control_entities=control_entities_config,
-        custom_batches=None,
-        sources=sources,
-        batch_size=batch_size,
-        fiscal_year=fiscal_year,
-        scope=scope,
-    )
+        kwargs["current_trace"] = current_trace
 
-    # Create list of date ranges
-    date_ranges = create_date_ranges(start_date, end_date, freq)
+    try:
+        # Extract entities for search querying
+        entity_keys = [entity.id for entity in companies]
 
-    no_queries = len(batched_query)
-    no_dates = len(date_ranges)
-    total_no = no_dates * no_queries
+    # Create entity configs
+        entities_config = EntitiesToSearch(companies=entity_keys)
+        
+        # If control_entities are provided, create a control EntityConfig
+        # For this example, assuming control_entities are all company entities
+        control_entities_config = None
+        if control_entities:
+            control_entities_config = EntitiesToSearch(**control_entities)
 
-    logger.info(f"About to run {total_no} queries")
-    logger.debug("Example Query:", batched_query[0])
-    # Run concurrent search
-    results = run_search(
-        batched_query,
-        date_ranges=date_ranges,
-        limit=document_limit,
-        scope=scope,
-        sortby=sort_by,
-        rerank_threshold=rerank_threshold,
-        **kwargs,
-    )
+        # Build batched queries
+        batched_query = build_batched_query(
+            sentences=sentences,
+            keywords=keywords,
+            entities=entities_config,
+            control_entities=control_entities_config,
+            custom_batches=None,
+            sources=sources,
+            batch_size=batch_size,
+            fiscal_year=fiscal_year,
+            scope=scope,
+        )
 
-    results, entities = filter_search_results(results)
+        # Create list of date ranges
+        date_ranges = create_date_ranges(start_date, end_date, freq)
 
-    # Determine whether to filter by companies based on document type
-    # For filings and transcripts, we don't need to filter as we use reporting entities
-    # For news, we need to check against our original universe of companies as a news article
-    # may mention other companies we're not interested in
-    needs_company_filtering = scope not in (
-        DocumentType.FILINGS,
-        DocumentType.TRANSCRIPTS,
-    )
+        no_queries = len(batched_query)
+        no_dates = len(date_ranges)
+        total_no = no_dates * no_queries
 
-    df_sentences = process_screener_search_results(
-        results=results,
-        entities=entities,
-        companies=companies if needs_company_filtering else None,
-        document_type=scope,
-    )
+        logger.info(f"About to run {total_no} queries")
+        logger.debug("Example Query:", batched_query[0])
+        # Run concurrent search
+        results = run_search(
+            batched_query,
+            date_ranges=date_ranges,
+            limit=document_limit,
+            scope=scope,
+            sortby=sort_by,
+            rerank_threshold=rerank_threshold,
+            **kwargs,
+        )
+
+        results, entities = filter_search_results(results)
+        # Filter entities to only include COMPANY entities
+        entities = filter_company_entities(entities)
+        
+        # Determine whether to filter by companies based on document type
+        # For filings and transcripts, we don't need to filter as we use reporting entities
+        # For news, we need to check against our original universe of companies as a news article
+        # may mention other companies we're not interested in
+        needs_company_filtering = scope not in (
+            DocumentType.FILINGS,
+            DocumentType.TRANSCRIPTS,
+        )
+
+        df_sentences = process_screener_search_results(
+            results=results,
+            entities=entities,
+            companies=companies if needs_company_filtering else None,
+            document_type=scope,
+        )
+    except Exception:
+        execution_result = "error"
+        raise
+    else:
+        execution_result = "success"
+    finally:
+        current_trace = kwargs.get("current_trace")
+        if current_trace and current_trace.event_name == TraceEventNames.COMPANY_SEARCH:
+            current_trace.workflow_end_date = Trace.get_time_now()
+            current_trace.result = execution_result  # noqa
+            send_trace(bigdata_connection(), current_trace)
+
     return df_sentences
-
-
-def collect_entity_keys(results: List[Document]) -> List[str]:
-    """
-    Collect all entity keys from the search results.
-
-    Args:
-        results (List[Document]): A list of search results.
-    Returns:
-        List[str]: A list of entity keys in all the search results.
-    """
-    entity_keys = set(
-        entity.key
-        for result in results
-        for chunk in result.chunks
-        for entity in chunk.entities
-        if entity.query_type == QueryType.ENTITY
-    )
-    entity_keys = list(entity_keys)
-    return entity_keys
-
 
 def filter_company_entities(
     entities: List[ListQueryComponent],
@@ -200,100 +204,6 @@ def filter_company_entities(
         for entity in entities
         if hasattr(entity, "entity_type") and getattr(entity, "entity_type") == "COMP"
     ]
-
-
-def filter_search_results(
-    results: List[List[Document]],
-) -> Tuple[List[Document], List[ListQueryComponent]]:
-    """
-    Postprocess the search results to filter only COMPANY entities.
-
-    Args:
-        results (List[List[Document]]): A list of search results, as returned by
-            the function `bigdata_research_tools.search.run_search` with the
-            parameter `only_results` set to True
-    Returns:
-        Tuple[List[Document], List[ListQueryComponent]]: A tuple of the filtered
-            search results and the entities.
-    """
-    # Flatten the list of result lists
-    results = list(chain.from_iterable(results))
-    # Collect all entities in the chunks
-    entity_keys = collect_entity_keys(results)
-    # Look up the entities using Knowledge Graph
-    entities = look_up_entities_binary_search(entity_keys)
-
-    # Filter only COMPANY Entities
-    entities = filter_company_entities(entities)
-    return results, entities
-
-
-def look_up_entities_binary_search(
-    entity_keys: List[str], max_batch_size: int = 50
-) -> List[ListQueryComponent]:
-    """
-    Look up entities using the Bigdata Knowledge Graph in a binary search manner.
-
-    Args:
-        entity_keys (List[str]): A list of entity keys to look up.
-        max_batch_size (int): The maximum batch size for each lookup.
-    Returns:
-        List[ListQueryComponent]: A list of entities.
-    """
-    bigdata = bigdata_connection()
-
-    entities = []
-    non_entities = []
-
-    def depth_first_search(batch: List[str]) -> None:
-        """
-        Recursively lookup entities in a depth-first search manner.
-
-        Args:
-            batch (List[str]): A batch of entity keys to lookup.
-
-        Returns:
-            None. The function updates the inner `entities`
-                and `non_entities` lists.
-        """
-        non_entity_key_pattern = r"\b[A-Z0-9]{6}(?=\.COMP\.entityType)"
-
-        try:
-            batch_lookup = bigdata.knowledge_graph.get_entities(batch)
-            entities.extend(batch_lookup)
-        except ValidationError as e:
-            non_entities_found = findall(non_entity_key_pattern, str(e))
-            non_entities.extend(non_entities_found)
-            batch_refined = [key for key in batch if key not in non_entities]
-            depth_first_search(batch_refined)
-        except (JSONDecodeError, RequestMaxLimitExceeds):
-            sleep(5)
-            if len(batch) == 1:
-                non_entities.extend(batch)
-            else:
-                mid = len(batch) // 2
-                depth_first_search(batch[:mid])  # First half
-                depth_first_search(batch[mid:])  # Second half
-        except Exception as e:
-            logger.error(
-                f"Error in batch {batch}\n"
-                f"{e.__class__.__module__}.{e.__class__.__name__}: "
-                f"{str(e)}.\nRetrying..."
-            )
-            sleep(60)  # Wait for a minute
-            depth_first_search(batch)
-
-    logger.debug(f"Split into batches of {max_batch_size} entities")
-    for batch_ in range(0, len(entity_keys), max_batch_size):
-        depth_first_search(entity_keys[batch_ : batch_ + max_batch_size])
-
-    # Deduplicate
-    entities = list(
-        {entity.id: entity for entity in entities if hasattr(entity, "id")}.values()
-    )
-
-    return entities
-
 
 def process_screener_search_results(
     results: List[Document],
@@ -405,8 +315,8 @@ def process_screener_search_results(
 
                     if not entity_key:
                         continue  # Skip if entity is not found
-
-                    # if entity isn't in our original watchlist, skip
+                    
+                    # # if entity isn't in our original watchlist, skip
                     if companies and entity_key not in companies:
                         continue
 
@@ -448,13 +358,12 @@ def process_screener_search_results(
         subset=["timestamp_utc", "document_id", "text", "entity_id"]
     )
 
-    df = mask_sentences(df, document_type)
+    df = mask_sentences(df)
     return df.reset_index(drop=True)
 
 
 def mask_sentences(
-    df: DataFrame,
-    document_type: DocumentType,
+    df: DataFrame
 ) -> DataFrame:
     """
     Mask the target entity and other entities in the text.